+++ include/muzzley/exceptions/AssertionException.h
@@ -42,8 +42,8 @@ namespace muzzley {
 			AssertionException(string _what,  int _http_code, int _code, string _desc, int _line, string _file);
 			virtual ~AssertionException() throw();
 
-			const char* what();
-			const char* description();
+			virtual const char* what() const throw();
+			virtual const char* description();
 			int code();
 			int status();
 	};