+++ src/exceptions/ClosedException.cpp
@@ -25,6 +25,6 @@ muzzley::ClosedException::ClosedException(string _in) : __what(_in){
 muzzley::ClosedException::~ClosedException() throw() {
 }
 
-const char* muzzley::ClosedException::what() {
+const char* muzzley::ClosedException::what() const throw() {
 	return this->__what.data();
 }