+++ src/exceptions/InterruptedException.cpp
@@ -25,6 +25,6 @@ muzzley::InterruptedException::InterruptedException(string _in) : __what(_in){
 muzzley::InterruptedException::~InterruptedException() throw() {
 }
 
-const char* muzzley::InterruptedException::what() {
+const char* muzzley::InterruptedException::what() const throw() {
 	return this->__what.data();
 }